core/requirements.txt
requests
python-dotenv
supabase
tqdm
solders
solana
tweepy